requirements.txt
asgiref==3.8.1
cached-property==1.5.2
certifi==2024.8.30
charset-normalizer==2.0.12
click==8.0.3
Django==5.1.1
django-cors-headers==4.4.0
django-js-asset==2.2.0
django-mptt==0.16.0
django-resized==1.0.2
expiringdict==1.2.1
gTTS==2.2.3
idna==3.8
mailru-im-bot==0.0.21
maxminddb==2.6.2
monotonic==1.6
numpy==2.1.1
opensearch-dsl==2.1.0
opensearch-py==2.3.1
pandas==2.2.2
pillow==10.4.0
python-baseconv==1.2.2
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.26.0
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==1.26.7
